Assistant checkpoint: Fix match scoring and display issues
Assistant generated file changes:
- shared/schema.ts: Change score columns to numeric type
- client/src/pages/Match.tsx: Fix match completion logic and pending scores display, Update points calculation, Fix display of match status

---

User prompt:

i can't end the game at AS again. it won't let the score be in .5 points. also, when the game is pending, i don't see the grey pending scores

+++ shared/schema.ts
@@ -84,8 +84,8 @@ export const rounds = pgTable("rounds", {
   startTime: text("start_time").notNull(),
   isComplete: boolean("is_complete").default(false),
   status: text("status"),
-  aviatorScore: integer("aviator_score"),
-  producerScore: integer("producer_score"),
+  aviatorScore: numeric("aviator_score"),
+  producerScore: numeric("producer_score"),
 });
 export const insertRoundSchema = createInsertSchema(rounds);
 export type InsertRound = z.infer<typeof insertRoundSchema>;
@@ -120,8 +120,8 @@ export type Team = typeof teams.$inferSelect;
 export const tournament = pgTable("tournament", {
   id: serial("id").primaryKey(),
   name: text("name").notNull(),
-  aviatorScore: integer("aviator_score"),
-  producerScore: integer("producer_score"),
+  aviatorScore: numeric("aviator_score"),
+  producerScore: numeric("producer_score"),
   pendingAviatorScore: integer("pending_aviator_score"),
   pendingProducerScore: integer("pending_producer_score"),
   year: integer("year").notNull(),
